Updated files

+++ src/roles/mixer.rs
@@ -6,7 +6,10 @@ pub fn run(args: &Cli) -> Result<()> {
     logging::println_tag(&tag, "starting (bridge MVP)");
 
     let sip = args.sip_bind.as_deref().unwrap_or("0.0.0.0:5063");
-    let target = args.target.as_deref().context("mixer requires --target (sink)")?;
+    let target = args
+        .target
+        .as_deref()
+        .context("mixer requires --target (sink)")?;
 
     // Preload config before UA init: codec modules + sip_listen so inbound leg actually listens
     let preload = format!(
@@ -16,7 +19,9 @@ pub fn run(args: &Cli) -> Result<()> {
          call_accept\tyes\n",
         sip
     );
-    unsafe { std::env::set_var("BRS_CONF_BUF", preload); }
+    unsafe {
+        std::env::set_var("BRS_CONF_BUF", preload);
+    }
     // Initialize baresip core/UA first (shared reactor/event bus)
     let ua = UaHandle::init().context("init UA")?;
     let (_bridge_ev, rx_ev) = ua.reactor.register_events();
@@ -27,13 +32,21 @@ pub fn run(args: &Cli) -> Result<()> {
             while let Ok(ev) = rx_ev.recv() {
                 let kind = format!("{:?}", ev.kind());
                 let text = ev.text.clone().unwrap_or_default();
-                if kind.contains("CALL_REMOTE_SDP") { logging::println_tag(&tag, "bevent: CALL_REMOTE_SDP (offer)"); }
-                else if kind.contains("CALL_LOCAL_SDP") { logging::println_tag(&tag, "bevent: CALL_LOCAL_SDP (answer)"); }
-                else if kind.contains("CALL_INCOMING") { logging::println_tag(&tag, &format!("bevent: CALL_INCOMING ({})", text)); }
-                else if kind.contains("CALL_PROGRESS") { logging::println_tag(&tag, "bevent: CALL_PROGRESS"); }
-                else if kind.contains("CALL_RTPESTAB") { logging::println_tag(&tag, "bevent: CALL_RTPESTAB"); }
-                else if kind.contains("CALL_ESTABLISHED") { logging::println_tag(&tag, &format!("bevent: CALL_ESTABLISHED ({})", text)); }
-                else if kind.contains("CALL_CLOSED") { logging::println_tag(&tag, &format!("bevent: CALL_CLOSED ({})", text)); }
+                if kind.contains("CALL_REMOTE_SDP") {
+                    logging::println_tag(&tag, "bevent: CALL_REMOTE_SDP (offer)");
+                } else if kind.contains("CALL_LOCAL_SDP") {
+                    logging::println_tag(&tag, "bevent: CALL_LOCAL_SDP (answer)");
+                } else if kind.contains("CALL_INCOMING") {
+                    logging::println_tag(&tag, &format!("bevent: CALL_INCOMING ({})", text));
+                } else if kind.contains("CALL_PROGRESS") {
+                    logging::println_tag(&tag, "bevent: CALL_PROGRESS");
+                } else if kind.contains("CALL_RTPESTAB") {
+                    logging::println_tag(&tag, "bevent: CALL_RTPESTAB");
+                } else if kind.contains("CALL_ESTABLISHED") {
+                    logging::println_tag(&tag, &format!("bevent: CALL_ESTABLISHED ({})", text));
+                } else if kind.contains("CALL_CLOSED") {
+                    logging::println_tag(&tag, &format!("bevent: CALL_CLOSED ({})", text));
+                }
             }
         });
     }
@@ -47,8 +60,15 @@ pub fn run(args: &Cli) -> Result<()> {
     // Start mixer legs: inbound catch-all and outbound dial
     sip_shim::mixer_init(sip, target, 8000, 1, args.ptime_ms)?;
     // Configure mixing: dtmf sequence and gains
-    sip_shim::mixer_config(&args.dtmf_seq, args.dtmf_period_ms, args.mix_gain_in as f32, args.mix_gain_dtmf as f32)?;
-    unsafe { std::env::remove_var("BRS_CONF_BUF"); }
+    sip_shim::mixer_config(
+        &args.dtmf_seq,
+        args.dtmf_period_ms,
+        args.mix_gain_in,
+        args.mix_gain_dtmf,
+    )?;
+    unsafe {
+        std::env::remove_var("BRS_CONF_BUF");
+    }
 
     // Event loop: wait for outbound (to Sink) established, then READY
     // Give outbound leg up to 5s to establish
@@ -67,6 +87,8 @@ pub fn run(args: &Cli) -> Result<()> {
 
 fn wait_for_ctrl_c() {
     let (tx, rx) = std::sync::mpsc::channel::<()>();
-    let _ = ctrlc::set_handler(move || { let _ = tx.send(()); });
+    let _ = ctrlc::set_handler(move || {
+        let _ = tx.send(());
+    });
     let _ = rx.recv();
 }
